+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   be placed on multiple entities)
 - Unload levels by removing the `YoleckKeepLevel` component from the entity
   that was used to load the level - or by despawning that entity entirely.
+- `YoleckSchedule::LevelLoaded` schedule for interfering with levels before
+  populating their entities.
 
 ### Change
 - `YoleckBelongsToLevel` now points to a level entity.